requirements.txt
pytest>=8.0.0
pytest-cov>=4.1.0
apache-beam[gcp]>=2.52.0
google-cloud-bigquery>=3.11.4